[IOS] style: Utility 그룹 생성
- manager 관련 파일 이동
